% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Funs.R
\name{add}
\alias{add}
\title{Sum of two vectors}
\usage{
add(x, y, plot = FALSE)
}
\arguments{
\item{x, y}{\code{Numeric} dqdzdzqd}

\item{plot}{\code{Logical} parameter. If \code{TRUE} the plot of \code{y~x} will be shown. Default to \code{FALSE}}
}
\value{
\code{Numeric} vector of \code{x+y}.
}
\description{
This \strong{wonderful} function will calculate for you the \emph{sum of any two numerical vectors}.
}
\details{
The method used to get \code{x+y} is obvious. No details are needed! The plot also is very  basic.

The user can provide any two vectors \code{x} and \code{y} and get \code{x+y} and the plot of \code{y~x}.

For more details see this \href{https://en.wikipedia.org/wiki/Summation}{article from Wikipedia}.
}
\examples{
# sum of two numbers
add(1, 1)

# sum of two vectors and a plot
add(c(1,2),c(1,8), plot = TRUE)

}
