% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my1pkg-package.R
\docType{package}
\name{my1pkg-package}
\alias{my1pkg}
\alias{my1pkg-package}
\title{my1pkg: My Personal First Package}
\description{
This is my personal package. It contains only one function.    
    The objective of this toy package is to see who to construct an R package.
}
\author{
\strong{Maintainer}: Anouar El Ghouch \email{anouar.elghouch@uclouvain.be}

}
\keyword{internal}
