% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Funs.R
\name{add}
\alias{add}
\title{The sum !}
\usage{
add(x, y, plot = FALSE)
}
\arguments{
\item{x, y}{\code{Numeric} vectors.}

\item{plot}{\code{Logical} telling whether to plot ggplot2 graph or not.}
}
\value{
\code{Numeric} vector of \code{x}, \code{y}, and \code{x+y}.
}
\description{
This \strong{wonderful} function will calculate for you the sum of any two numerical vectors.
}
\details{
The calculation of \eqn{x+y} is obvious. No details are needed!
}
\examples{
# sum of two numbers
add(1, 1)

# sum of two vectors
add(c(1, 2), 1:8)
}
\seealso{
For more details see \url{https://en.wikipedia.org/wiki/Summation}.
}
